Improve image quality
Currently, image quality on the website is a little wrong. This is mostly because we convert stuff to PNG. This commit fixes this.

Note that this commit was written while drunk and was not tested at the time of writing. Will test as soon as I think about this.

Why we do this? Look at https://gewis.nl/photo/view/121342, and note the MWM logo, you see some kind of gradient in the logo here. Now look at the source image https://gewis.nl/photo/download/121342 where we definitely do not see any kind of gradient.

Also, note that serving files as PNG while they are originally JPEG, will actually transfer a lot of data which we do not need. PNG is a horrible format when it comes to serving photos, in which case lossy compression is nice (PNG is great however for things where we need lossless compression.).