Use different subject for member initiatives
A generic hard-coded text is used instead of the user's name to clearly distinguish between organ activities and member initiatives. Maybe there's a member that happens to be called Web ;) Note that no translation is applied on the hard-coded text. This is done on purpose because translating the email's subject seems overkill.