Toggle metadata when pressing info button
The PhotoSwipe library hides unknown buttons by default. The info button can be enabled by explicitly setting the infoEl to true in the options. However, to invoke the same behaviour as the share button (i.e. toggle the modal), we need to make some changes to the library's code. Unfortunately, it isn't possible to pass a custom callback.