Display job creation date in a human readable way
I changed the signature of the setTimeStamp() method to only accept instances of DateTime. The only place where this method is called is in the Company service where it is passed a new instance of DateTime. Changing the signature ensures that no type checks are needed before calling Carbon::instance().