Show 'today' if the job is posted today
The table only stores the date without the time. So calling Carbon's diffForHumans() method on that date will assume the time to be midnight. For dates in the past this doesn't matter because Carbon will display 'X days ago'. However, for job's posted on the same date as a user's visit, Carbon will calculate hour differences. This doesn't make sense.

Note that returning a Carbon instance instead of a DateTime instance in the getTimestamp() method is safe because Carbon extends DateTime.