Removed Grunt and added node-sass
When creating a self-contained docker container for the website (needed
for migration to PHP 7 and keycloak), we had some issues to compile the
CSS, because ruby-sass is relatively complicated to install.

Since the entire mess was quite complicated anyways, having to install
npm, grunt and ruby-sass, we think it is better to move to node-sass
instead. This required renaming the main.scss to gewis-theme.scss, which
isn't really a big change.

So, the new procedure for compiling the scss is as follows:

- Have npm installed
- Execute `npm install`
- Execute `npm run scss`
- ???
- Profit!

For ease during development, we added the command `npm run watch`, which
is equivalent to the `grunt watch` command.