Add button to switch themes
Adds a button to the main navigation bar which is used to switch themes. It also showns the current theme.

In `layout.html` I introduced a theme switching script based on Derek Kedziora's Dark Mode Revisisted. The script is inline as this prevents issues with loading the script later. We need to have the theme determined as soon as possible (i.e., after the DOM has finished loading).

To use the theme across multiple tabs and for longer sessions (specifically for the user overwrite) the preferred theme is stored in `localStorage`.

As `localStorage` is not always and everywhere available, the theme cannot be switched by the user when this it is not possible to store this information.

Furthermore, in the event that the system preferred theme cannot be determined (e.g., due to missing support) the user can still switch themes.

By default the day theme is applied.