Resolve input validation errors when updating or creating an activity
This resolves an issue that would lead to errors when the activity creation form is submitted or the activity editing form is viewed.

This is because certain values would be strings (i.e. directly from the form), however, the code expects a `LocalisedText` object.

This is resolved by adding an additional check for strings.