Introducing Acitivity SignupLists! Activities are no longer activities, they are now Activities and SignupLists. With this change there are a lot of new possibilities for those who organise activities.
Every Activity can have multiple SignupLists, this means that
separate Activities for things such as food subscriptions are no
longer necessary! Now you can just add an additional SignupList
to your activity and you have a food subscription list at your
disposal.

Each SignupList has a name, an open date and time, and a closing
date and time. Furthermore, the old 'additional fields' can now
be used in a SignupList. Want to get some more information from
the participants? There are (still) several options available:
    - text
    - number (range)
    - yes/no
    - choice

There are currently some limitations with this change. Updating
activities is currently not supported. In addition to this, there
is still legacy code in the depths of the changed files.