Hide companies and jobs when they should be hidden
Companies which were hidden were not actually hidden. And jobs for
hidden companies were still visible. This adds a new condition to
the company show action. Now, companies and jobs are only shown when
they should be shown.

To accomodate for this change, the boolean expression that determined
whether or not a company should be hidden is changed. To explain the
changes the truth table is given below. v is the `$visible` variable
and h is the value of `getHidden()`.

v h | isHidden
0 0 |    1
0 1 |    1
1 0 |    0
1 1 |    1

This results in isHidden := !v || h. In other words, explicitely
setting the hidden property of a company overwrites whether or not
the company should be visible (i.e., not hidden).

For Jobs this meant the small addition of a negation to make a job
inactive when the company is hidden.