Introducing Activity Categories
Gotta catch em' all!

This update enables the creation of activity categories for admins. After creation, the categories can be edited and deleted.

When creating activities, you are now greeted with a new option to add activity categories to your activity.

When viewing approved activities the new activity categories are now displayed.