Allow more Activity attributes to be updated This commit introduces new functionality to improve the Activity updating experience. Instead of the basic Activity attributes (e.g., name, location, and description), you can now also update the start / end date and time, the organising organ, the organising company, and the activity categories.
Updating SignupLists is not allowed, as this can possibly
involve creating new data points for an activity, which are not
easily transferred from the UpdateProposal's $new to the $old
Activity.

Furthermore, the time an Activity can be updated has been
reduced. Activities can only be updated before they begin or
the opening date and time of the earliest SignupList.

Duplication between the creation form and the update form has
been removed.