Allow SignupLists to be edited. Fixes #728. This change allows SignupLists to be edited after an Activity has been created. However, there are some restrictions to this functionality. Activities which are already approved and have one or more SignupLists can no longer be edited (a warning is shown).
Activities which are already approved but do not have any
SignupLists can still be edited, however, it is not possible to
add new SignupLists to this activity.

Previously old an old Activity would receive the data from the
new Activity in the ActivityUpdateProposal. However, with the
introduction of SignupLists this has become too complicated.
Therefore, the old Activity will now be disregarded. This means
that an updated Activity will have a new id.

Furthermore, when updating an Activity this would previously
trigger the standard "An activity has been created" e-mail. This
is no longer the case, there is now an e-mail which says an
update has been made and another e-mail which says there is a
new update proposal which needs to be approved/dissaproved.

A dissaproved Activity can still be updated, however, it will
now always create an update proposal (instead of directly
applying the update). This ensures that dissaproved activities
get another chance. Furthermore, when the update proposal of
a dissaproved Activity is approved, the updated Activity will
be moved to the "not yet approved"-queue instead of remaining
dissaproved.

To resolve an issue with Zend's form validation, the option
"Allow external participants" has been renamed to "Only allow
GEWIS members", which is checked by default.