Fix issues preventing updating activity with choice option
Multiple issues prevented users from updating an activity if it
contained a choice field in its sign-up lists:

1) Missing data in update form. An issue existed where the language
specific values of a choice option were not present in the update
form. This was caused by incorrectly serializing fields as arrays.

2) Data structure parity inconsistencies. Form submission and known
activity data structures were different for the choice option on
sign-up lists. Existing activities represent the data as an array
while the form represents the data as a single string.

Both issues were resolved by better handling of the underlying
data.

3) An issue existed in the SignupField model where SignupOptions
could not be removed from the database due to a missing
orphanRemoval tag.