Create feed and edit pages
Create feed for the vacancies section with redirects to edit pages for
each vacancy in the feed. Also created action for the edit pages and set
the values in them based on the jobs pulled from the database. Loading
into the database doesn't work yet for central values as they are not
bound to the form.