Fix role order to ensure correct privileges for each user The Laminas ACL changed the order in which inherited roles are applied. Previously, all roles would be checked, however, this is no longer the case. Now roles are checked LIFO (last-in-first-out), which means that the most privileged role should be applied last.
This was not the case, which meant that users who should get most privileges (`company_admin`s, `admin`s, and board members) did not get the correct privileges.

To prevent this from happening, we devised a cleaner solution which directly applies the role of a User in the correct order. The order of most privileged to least priviliged is defined as follows:
- `admin` / `board`
- `company_admin`
- `active_member`
- `user`
- `photo_guest`

This means we have now also introduced a new role for board members (`board`), this is because the previously existing assertion does not correctly function when only applied to the `user` role (as board members are almost always `active_member` or `company_admin`). Applying the assertion to all roles caused other issues with the all privileged `admin` role. Since the role of a board member is absolute, an assertion (which should be used to determine rules that are not absolute) is ill-suited for determining the privileges for a board member.

This fixes #1123.

Co-Authored-By: Koen Teuwen <32361020+Koen1999@users.noreply.github.com>