Remove never shown message for unsuccessful option creation
The form itself gives errors which detail what the user should do. It is therefore not necessary to have an additional message to let the user know the form is invalid. The `success` variable was removed in ea5431dd129ee8f146317c213ef2b1e64b3b279e in favour of custom error messages in the form.