Fix incorrect merge of Company data by Doctrine
Doctrine cannot perform merge operations on its own, the last operation
performed is what a UnitOfWork will execute. In this case it meant that
saving a company's information would not only not save any new
information regarding its I18n language profiles, the language
profiles would be proactively deleted.

To fix this, edits to language profiles are not first persisted and
after that has completed, any languages which are to be removed are
removed.

Removing a language profile now removes it from the database and not
random artefacts in the database.

This also includes a (temporary) fix for companies with only one
language, where the logo cannot be found if viewed with another locale
than the one the company's information is defined for.