Remove Hydrator strategy which breaks signing up to sign-up lists
The ClassMethodsHydrator is no longer compatible with the Signup
Form due to some magic constant. The hydration strategy does not
provide any benefits to our current use-case, hence removing it
does not hurt the website.

In the future I would recommend looking at
doctrine\doctrine-laminas-hydrator to see if that works better
than the standard Laminas Hydrators.