Move to new Supremum Infima API
The Supremum Infima API has moved to a new location and now requires
a key for authorisation. Using Laminas' cURL adapter, we retrieve an
infimum from the API, which is then served through our locally
authenticated API.

This ensures the client is always the requesting party and as such
we do not have a blocking request in every authenticated response.
Furthermore, we do not expose the API key to the client.