Reduce response time by bypassing Doctrine entity
Previously, Doctrine spent most of its time processing
the `Member` entities when retrieving the results from
the database. By moving to an entity-less query (e.g.,
HYDRATE_ARRAY) I was able to reduce the response time
by ~40%.

To further improve this, I have completely moved the
query to native SQL. Including the generation of the
full name. This decreased the response time another
~40% for a total of 60%.