Make it possibly to directly link to images
We do not strictly know the 'position' of a photo in an album,
hence directly linking with `#&gid=1&pid=[position]` is not
possible without determining the position of the photo.

Luckily we can use custom PIDs. I have opted for the easiest
solution, which just uses the id of the photo. Hence we can
use something like `#&gid=1&pid=42` to link to the photo with
`id=42`, however, we still need the album to complete the url.

`/photo/(album|member)/:album_id#&gid=1&pid=42`

If the id does not exist in the album (i.e., the photo with
that id is not part of the album), you directly view the
album.

One potential side-effect, manual enumeration is a lot more
difficult.