Fix all level 6 PHPStan issues
This is a squashed commit. Originally it were 8 commits, however,
due to the complexity and connections between the modules it has
been reduced.

PHPStan was run with following changed parameters:

```yaml
level: 6
treatPhpDocTypesAsCertain: true
checkMissingIterableValueType: false
checkGenericClassInNonGenericObjectType: false
```

in combination with a suppressor rule for the controler actions:

'#Method [a-zA-Z0-9_]+\\Controller\\[a-zA-Z0-9_]+\:\:[a-zA-Z0-9_]+Action\(\) has no return type specified.$#'

The original commit messages were as follows:

Fix all issues with the Activity module

Fix all issues with the Application module

Fix all issues with the Company module

Fix all issues with the Decision module

Fix all issues with the Education module

Fix all issues with the Frontpage module

Fix all issues with the Photo module

Fix all issues with the User module