Fix broken synchronous AJAX requests for templates in option calendar
Somewhere near the end of 2021, Chromium-based browsers appeared to have changed their behaviour regarding synchronous XMLHttpRequests (AJAX). This resulted in a race condition, where the templates would not be loaded before they were used.

This commit slightly modifies the calendar library (as it has not been updated in a few years, I do not foresee any future problems), to make the AJAX requests asynchronous and adapt the code to the new behaviour.