Replace old activity option calendar with new upgraded alternative
This replaces the old bootstrap calendar with FullCalendar, very similar
but also not. Comes with a lot of changes:

- No longer dependant on the successful retrieval of the pre-planned
activities in the Google Calendar, proposed options will always be shown.

- Better monthly view, it is easier to see the duration of activities
which span multiple days.

- Improved day view for approval of proposed options. Furthermore, if an
option has been approved, the approval button is no longer shown.

- Less dependencies; jQuery and underscore.js are no longer required.

Unfortunately, also a small caveat, while Bootstrap is supported as a
theme, our v3.4 version is not compatible. Hence, internal styles and
icons are used. This means a changed CSP (see NGINX config).