Fix responsiveness panorama
Unfortunately, `image-set` does not support a `width` attribute. This means that it does not really have a purpose for us. It has been replaced with an actual `<img>` by utilising `srcset` (which does accept widths).

The images have been resized and compressed using `cwebp` with `-q 90`. This produces acceptable file sizes which still have good image quality. There are many options to choose from, so the user will always see something which looks good on their device.

The CSS contains a small hack to get the `inset` shadow back, but it is not super complicated. Also fixes an issue where the panoramo used its true center and not 40% from the top.