Fix ordering of activity archive
Order of the navpills is not the same as in the photo module, which
is more logical (decreasing order). Also displays the activities in
reverse order (most recent is first).

Also some small quality of life improvements, consisting of small
messages if there are no activities/albums in a particular year.