Load tags dynamically via API
This introduces a simple API which can return tag information for
a certain photo. This was a bottleneck for large albums, as
Doctrine would load all information about tags (which includes
the associated entities).

Similarly to the search API, only the required fields are returned
from the database to prevent us from loading complete entities.

When the photo viewer is opened, it will (try to) get tags from
the API whenever a new slide is preloaded. This means that
currently, on the initial click, tag information from 3 slides
is loaded. After that, it will only load data for 1 slide at a
time and only for slides which have not yet been visited.