Show actual error page instead of blank page for unrecoverable PHP errors
To make sure the user has a better experience when an unrecoverable
(and thus fatal) `Throwable` is not caught I have introduced a global
error page.

This makes sure that users actually get to see that something went
wrong, instead of only getting a blank page. The page will only be
shown for `Throwable`s that could not be caught by Laminas MVC (e.g.,
`PDOException`s).

Additionally, I added the page (and NGINX configurations) that I use
for maintenance periods in which NGINX is still online, but the
actual application not.