Actually log the fatal error and set status code for Zabbix
To make sure that we can actually see what is going on, we must
not forget to actually log the error somewhere. It will be logged
to `php.log`.

Additionally, by using `readfile` the response status code is
automatically 200. This breaks the uptime reporting provided to
us by the CBC. Hence, we force the status code to be 500.