Add second endpoint to replace `HasVoted` view helper
Just like the `tags` endpoint, this endpoint is used to determine
whether someone has voted for a certain photo.

I have tried to make only make one request containing both the tags
and vote status, however, due to uncertainty in the timing of
events in PhotoSwipe this does not work. Nor can the `fetch`
promise be split into two separate promises to be resolved like we
do now.