Combine tags and voted requests into one request
After a whole lot of trying things and failing miserably I have
managed to get the one request for everything working.

At `gettingData`, PhotoSwipe will try to request the "details" of
the photo(s) it is loading (only if for that particular photo
there are no tags in the actual DOM, we do not check for the voted
status as we can assume it will not exist either). Subsquently,
when `change` is signalled, the request will be resolved (async).
Once that is finished, we dispatch our own events to update the UI.

This is the only way this can be done, without dispatching events
it is not possible to update both UI elements with one request, it
does not matter at which stage you are actually loading the data.

Directly resolving the request also does not work, as there is no
good way to ensure that the data is still present when the slide
or its contents are shown.