Add scoped claims to JWT authentication
This adds scoped claims to the JWTs we use to let external apps
authenticate users.

An additional attribute has been added to the `ApiApp` model, aptly
called `claims`. This attribute (in the database) is a comma-
separated list of claims that can be found in the `JWTClaims` enum.

The available claims (currently) are:

- `email`,
- `family_name`,
- `given_name`,
- `is_18_plus`,
- `lidnr` (did technically already exist), and
- `membership_type`.

Retrieving the actual values is also handled from within the enum.

This also fixes an issue where the website would crash if an invalid
`appId` was provided.