Add authorisation prompt when authenticating external application
This adds a prompt into the JWT authentication process, which makes
it a proper authorisation by the user.

The prompt will show the claims the external application will
receive. If the user has previously authenticated with an application
and that was less than 90 days ago, the user will not be prompted.
However, if this was more than 90 days ago, the user will receive
a reminder (cannot be dismissed) as to what claims the external
application has access to.

In the event that the user does not agree, they will be redirected
back to the application using a newly added "application URL".

First and the most recent authentication for each external
application can be found on the member's personal page.