Fix typing/style issues
There are still two errors in PHPStan, however, Doctrine works this
way. `$claims` cannot be `null` because it is not backed by the enum
as a result ORM will return an empty array `[]`.

If there are other unknown values, ORM will throw an exception as it
cannot convert the value to an enum. There were talks about so called
`fallbackEnumValue`s in ORM, however, that proved to cause only more
problems and hence has not been merged (and probably never will).