Do not translate organ member function before sorting
This would have worked if `Voorzitter` was also translated, but that
was not the case. It would have been possible to do this, but then
the translator is called on every sorting which is not particularly
efficient.

To fix this, we just use the Dutch term for sorting and translate
the values in the views.