Make Photos of the Week a `VirtualAlbum` per year
This adds another `VirtualAlbum`: `WeeklyAlbum`. This album is used
to hold all the Photos of the Week in a certain association year,
it has an extra attributes `$dates` that holds an array of the
`DateTime`s for each Photo of the Week.

Because the photos are now in a `VirtualAlbum`, we can reuse the
existing photo viewer to show the album. So each year of Photos of
the Week can be viewed like any other album.

In the photo viewer, if a photo is a Photo of the Week a small
box (like the tag list box) is shown to let the user know in which
week of what year the photo was the Photo of the Week.

The Photos of the Week overview page has been revamped, it still
shows the current Photo of the Week as it did before, but it no
longer lists all Photos of the Week separately. Now, you can see
all association years (the thumbnail of each "album") is the last
Photo of the Week for that year.