Automatically calculate aspect ration on persist
This automatically calculates the aspect ratio of a `Photo` when
the entity is persisted (only on the first persist). There is no
other way to do this without involving the EntityManager, as
Doctrine is strictly follows the data mapper pattern and not the
active record pattern. Hence, updating the aspect ratio from
within the entity is not possible.

This does still require the manual calculation of the aspect
ratios of all old photos.