Make member e-mail address nullable and change registration and password reset forms
This ensures that we are on the same level as gewisdb. This means
that member's e-mail addresses can now be nullable. This requiered
some changes to the existing forms.

The registration form no longer asks for an e-mail address, members
without an e-mail address cannot register for an account on the
website.

The password reset form still requires an e-mail address to be
submitted, because we assume they will know it.