Remove and clean-up legacy services
This removes the `Legacy` service which has not been in use since
the migration to Docker.

It also removes the old pin login, as
that has not worked for quite some time too. This simplifies the
`LoginAttempt` code a lot.

Lastly, an old API endpoint has been removed that exposed too much
data and has not been used (was previously used by REX).