Fix CSP violation and broken redirect on authorisation of ext. app.
This fix has gone through a few iterations, which all have proven to be
unable to fix this issue. For documentation purposes, I believe it is
important to still highlight the changes I have tried:

0) Splitting GET and POST action: this step was not intended to fix the
issue, but as the router supports `Method` routes, I tried to split the
functionality to make the authorisation process clearer.

1) Adding additional GET after POST: I implemented an additional GET
request after POSTing the authorisation. This could either be a "grant"
or a "decline". This does not work, as the CSP code in Chromium is
smart enough to recognize this.

2) Adding a manual `http-equiv` refresh: literally this, altough this
did not work with the `Location: ` based redirect still in place.

Ultimately, I have reverted most of the code and just added the
simplest redirect template. However, I used the `layout` plugin to
change which template was being used. This was stupid, as it broke how
the website looked (frontpage layout was overwritten) and did not add
the refresh `meta` tag.

Using the `ViewModel`'s `setTemplate` did work as expected.