Fix PHP-FPM config not being used
Unfortunately, the default configuration of PHP-FPM in our Docker
containers always changed our configurations. This is because the
default config is called `www.conf` and ours `php-fpm.conf`. From
inspecting the container, it is evident that the Docker config is
called `zz-docker.conf` to ensure that the config is correctly
changed.

Hence, to ensure our config is also used correctly, it will be
named `zz-{gewisweb|glide}.conf` for the website and Glide
respectively.

I have also slightly tweaked the config itself to make a bit more
sense.