Switch to Alpine docker images
This changes the base images for our Docker containers to Alpine,
this reduces the size and hopefully makes it easier to work with
them.

Furthermore, I have updated the NodeJS images in the production
version of `web` to LTS. This ensure that we both use LTS during
development and production (dev was already LTS).

Matomo is not yet updated, this requires more changes (changing
to PHP-FPM).