Move Matomo to PHP-FPM and use existing NGINX setup as reverse proxy
This fixes the version of Matomo (in this case to v4.10+, as the
images for v4.11 have not yet been released), and it changes from
the Apache variant to PHP-FPM Alpine.

To accommodate for this change, the NGINX configuration has been
updated with the best practices for Matomo. A new shared volume has
been introduced to ensure that static files can be served without
having to go through PHP-FPM.