Change CSP to require nonces for scripts
This change requires that any script embedded on the website
has a `nonce`-attribute. An environment variable has been made
available (`NONCE_REPLACEMENT_STRING`), which is supposed to
contain a cryptographically secure random string which NGINX
will replace with the actual nonce.

Unfortunately, as we act as a reverse proxy, we do not handle
SSL/TLS connections, hence my proposed `$ssl_session_id` is
empty and does not contain a cryptographically secure random
"nonce". Because OpenSSL is enabled, the `$request_id`
variable is a valid replacement (also secure).

In the HTML template files, the constant
`NONCE_REPLACEMENT_STRING` can be used to put the placeholder
into the output, which is subsequently replaced by NGINX.

Stylesheets are excluded from from this change, as I strongly
believe that they do not pose any threat. See this blog post
by Scott Helme for more information:
https://scotthelme.co.uk/can-you-get-pwned-with-css/.