Fix localisation issues
It appears that Alpine Linux does not come with full localisation
support by default. Any reasonable information on this is actually
pretty sparse, so it took me more than an hour to fix this.

I initially tried to use `musl-locales` and `musl-locales-lang`,
however, while these provide partial localisations (for Dutch) PHP
would not accept it no matter what. Even moving the files around
didn't fix the issue.

Finally, I saw https://github.com/docker-library/php/issues/1302
which explained that `icu-data-full` is not installed by default,
as it is approximately 30MB (which is quite large for Alpine).