Replace CKEditor 4 with CKEditor 5
CKEditor 4 is not compatible with our new CSP, it requires
`unsafe-eval` and `unsafe-inline`.

The build of CKEditor 5 is a custom build tailored to the GEWIS
website. It is available from
https://github.com/GEWIS/gewisweb-ckeditor.

However, there are some downsides to this. "General HTML Support",
CKEditor 5's content filtering and modification plugin does not
support classes on `ul`s and `ol`s. They are downcasted to the
`li`s inside. This breaks the nav pills that are used on some
pages. Ideally, we update our CSS to allow the usage of `div`s
instead of `ul` and `li`. In contrast, it is a lot better at
filtering content than the equivalent in CKEditor 4.

Additionally, `URLHelper` does not work with the editor. The
upload endpoint is now hardcoded in the editor source code (see
repository).