Improve performance of `/career` page by reducing #queries
This takes advantage of a customised native query to select all
public companies (which are companies that are published AND have
at least one active package).

Furthermore, there were unnecessary calls to get all job categories
that caused a lot of additional queries. Furthermore, improved
logic allows us to calculate the number of active jobs in a
category to be calculated only once instead of three times.