Allow authorizations to be revoked
This introduces new functionality that allows authorizers to revoke
their authorization for a GMM.

This also improves the admin overview for authorizations, by removing
the notion of "invalid" authorizations and splitting valid
authorizations and revoked authorizations. Both types of
authorizations contain timestamps to determine when the authorization
was made and revoked.