Fix slow or no download for exams and summaries
This changes the backend for generating and applying the watermarks
to PDFs. ImageMagick has been replaced with FPDI and TCPDF.

Furthermore, the PDFs are no longer flattened (the real cause of
the problem) but to ensure that the watermark cannot be removed a
random password (which is subsequently encrypted with AES-256) is
applied to the PDF ensuring that it can only be printed. By not
flattening the PDF, the experience of using the PDFs is also
improved as it allows for the selection of / searching for text.