Try to recursively generate a photo album cover
This replaces the original implementation for generating photo album
covers by a recursive solution.

The idea is that (to a certain depth) sub-albums and their contents
are retrieved for an album and the first level of full albums will
be used for the cover.

Unfortunately, MySQL/MariaDB's recursive CTE implementation is limited,
so an early abort is not possible. For large albums this means that the
query can be quite heavy (in the order of 0.02s).