Allow for the creation option proposals in a certain period
This adds an additional selector to the option proposal form that allows
the user to select a certain period in which the options are placed.

This does not influence the behavior for admins and board members, they
can still propose options regardless of a period.

This also fixes the form validation on the options fieldset.