Give `tueguest` to IP addresses in one of multiple ranges
The university has started to roll out a new NAT to free up some of
the IP addresses in 131.155.0.0/16. This breaks the `tueguest`
privilege that should be assigned to everyone in the TU/e network.

This change addresses that, by making the check work for multiple
subnets. Including the new 10.64.0.0/10 NAT subnet.