Add traits to make it easier to work with repetitive tasks in entities
This adds a few traits which should make it easier to do certain
things with entities.

There is a new `ApprovableTrait` which aims to make approving and
rejecting of entities streamlined. It provides several columns to
entities using this trait:

- `approved` (ApprovalStatus enum)
- `approvedAt` (`DateTime` at which the entity was approved)
- `approver` (the member who approved the entity)

The `ApprovalStatus` enum makes use of PHP 8.1's enum type to
define several state in which an entity can be; unapproved,
approved, or rejected. Currently, this enum is not natively used
by Doctrine, as this is expected to be implemented in ORM v2.11.

The next trait, the `IdentifiableTrait`, is just there for the
`id` column of an entity. Should reduce a lot of repetition and
incompleteness with `getId()` and `setId()`. At this point in
time, not all entities have been converted to use this trait.

Finally, the `TimestampableTrait`, which can be used to track
when an entity was created and/or updated. These values are
automatically updated by Doctrine at `persist()` or `flush()`.