Update Company ACL to account for complex resources and privileges
This update modernises the Company ACL to ensure that only certain
people can do certain things. This mainly focusses on the difference
between users (`company_admin`) and company users (`company`).

To properly destinguish between the different types of actions, for
example, creating companies or jobs, new resources have been added.
This should make it clearer who can do what with that.

Ideally, an assertion is added for `company_user` to ensure they
can only edit categories and jobs that belong to their company.

This also makes use of named parameters in the `allow()` functions
to make it a bit clearer what is going on, withouth having to
expand the arrays (that would make it an eyesore).