Add basic functionallity for proposing jobs as a company user
This allows `CompanyUser`s to propose new jobs. It re-uses the
existing job form from the admin interface (DRY).

Furthermore, there is a basic implementation of the jobs
overview page(s) for companies. This will require significant
updates to both the backend and frontend to make usable.

The idea is that on the main jobs page (/company/jobs) you can
see active and expired job packages. For the active job
packages you have the ability to add a job. Additionally,
there is an overview with the `n` most recent (un)approved/
rejected jobs and potentially notifications.

You can also inspect a job package (when active) in detail,
which provides an overview of *all* jobs related to that
package, not just `n` most recent jobs.