Allow initial job proposals to be (dis)approved
The `ModifyRequest` form previously used by the `Activity` module is
now a form in the `Application` module as we now also use it in the
`Company` module.

Similar to the `Activity` module, job proposals can now be viewed by
people with the proper privileges (`approve` for `job` in the ACL).
Here the user has the option to either approve or dissaprove the
job, an additional message can be left. However, that message is not
yet visible to companies.