Fix no uniqueness across representative e-mail addresses
This addresses an issue where multiple companies could have the
same e-mail address for a representative. This would lead to
newer companies in such a situation being unable to login or
being able to request a password reset for the old(er) company.

Note that actually resetting the password was never possible.