Change how jobs are displayed to companies
A small rework of the 'Jobs' overview for companies. When selecting a
specific package, all jobs are now displayed in a table instead of
separately based on their approval status.

Furthermore, the overview of job packages now also displays the 5 most
recent jobs for a specific approval status.

Added stubs for deleting jobs and transfering jobs from expired
packages.