Allow update proposals to jobs to be (un)approved
Adds a UI similar to the `Activity` update proposal diff UI that
show changes between an original `Job` and an updated version.

If the update is applied, the changes are migrated to the original
job and any older updates (e.g., rejected onces) are removed.

If the update is rejected, everything stays as it is.