Add Pwned Passwords checks when handling passwords
This adds functionality to check for passwords that have appeared
in public data breaches.

When activating an account or changing a password you will not be
able to use passwords that have appeared in public data breaches.

When logging in, you will be requested to change your password,
however, this is not a hard requirement (i.e., you can navigate
away from the "reset your password"-page).

Unfortunately, setting the `mustChangePassword` property on login
is difficult. We have to detach and "re-find" the user object to
be able to persist it. This issue is also present in the
`LoginAttempt` service.