Enforce change of password when pwned
Unlike having a "soft" enforcement of a pwned password as introduced
in 9e272a7a14add47b437750820531416fbd3de79c, this change requires
users to request a password reset before being able to log in when
their password is determined to be in a public data breach.

This make this possible, the Pwned Passwords API check was moved to
a separate service and the checks are now performed during the
authentication process. This ensures that if a password is pwned,
the user is not logged into the website.

If something fails during the login process, identifiers and
passwords are now removed from the stack traces using the new
`#[SensitiveParameter]` attribute.